[LCR-99] refactor: 페이지들 suffix Page로 수정 & 모든 파일 절대경로로 수정
