+++ src/apis/ItemCreatePage.ts
@@ -1,8 +1,10 @@
 import { api } from "@/apis/config/Axios";
-import { ItemRequest } from "@/types/api/ApiRequestType";
+import { ItemCreateRequest } from "@/types/api/ApiRequestType";
 import { ApiResponse, NoResponse } from "@/types/api/ApiResponseType";
 
-export const postItem = async (data: ItemRequest): ApiResponse<NoResponse> => {
+export const postItemCreate = async (
+  data: ItemCreateRequest,
+): ApiResponse<NoResponse> => {
   const response = await api.post("/items", data);
   return response.data;
 };
