+++ src/components/common/Modal.stories.tsx
@@ -1,5 +1,5 @@
+import Modal from "@/components/common/Modal";
 import { Meta, StoryObj } from "@storybook/react";
-import Modal from "./Modal";
 
 const meta: Meta<typeof Modal> = {
   title: "components/common/Modal",
