+++ src/hooks/useAuth.ts
@@ -1,7 +1,7 @@
 import { useAuthStore } from "@/stores/useAuthStore";
-import { useAuthApi } from "./api/useAuthApi";
 import { ErrorMessage } from "@/utils/ErrorMessage";
 import { LoginRequest } from "@/types/api/ApiRequestType";
+import { useAuthApi } from "@/hooks/api/useAuthApi";
 
 export const useAuth = () => {
   const { isLogin, setLogin, setLogout } = useAuthStore();
