+++ src/pages/onBoardingPage/views/OnBoardingContents.tsx
