+++ src/pages/popUpListPage/PopUpListPage.tsx
@@ -5,7 +5,7 @@ import { useNavigate } from "react-router-dom";
 import { Swiper, SwiperSlide } from "swiper/react";
 import { Navigation, Pagination } from "swiper/modules";
 import { PopUpCards } from "@/mocks/handlers/popUpList/PopUpCards";
-import PopUpCard from "@/pages/popUpList/views/PopUpCard";
+import PopUpCard from "@/pages/popUpListPage/views/PopUpCard";
 import Modal from "@/components/common/Modal";
 import logoImage from "@/assets/webps/common/logo-manager.webp";
 import plusWhite from "@/assets/webps/popUpList/plus-white.webp";
@@ -14,7 +14,7 @@ import rightArrowGray09 from "@/assets/webps/common/right-arrow-gray09.webp";
 import bin from "@/assets/webps/common/bin.webp";
 import check from "@/assets/webps/common/check.webp";
 
-export default function PopUpList() {
+export default function PopUpListPage() {
   const navigate = useNavigate();
   const [isConfirmModalOpen, setIsConfirmModalOpen] = useState(false);
   const [isAlertModalOpen, setIsAlertModalOpen] = useState(false);
