+++ src/pages/popUpListPage/views/PopUpCard.tsx
